+++ build/skc-tools/skc_library/deploy_scripts/deploy_skc_library.sh
@@ -3,6 +3,14 @@ SKCLIB_BIN=bin
 SGX_DRIVER_VERSION=1.36
 SGX_INSTALL_DIR=/opt/intel
 
+# Check OS and VERSION
+OS=$(cat /etc/os-release | grep ^ID= | cut -d'=' -f2)
+temp="${OS%\"}"
+temp="${temp#\"}"
+OS="$temp"
+VER=$(cat /etc/os-release | grep ^VERSION_ID | tr -d 'VERSION_ID="')
+OS_FLAVOUR="$OS""$VER"
+
 source skc_library.conf
 
 KDIR=/lib/modules/$(uname -r)/build
@@ -35,11 +43,24 @@ install_dcap_driver()
 
 install_psw_qgl()
 {
+
+if [ "$OS" == "rhel" ]
+then
+# RHEL
+
 	tar -xf $SKCLIB_BIN/sgx_rpm_local_repo.tgz
 	yum-config-manager --add-repo file://$PWD/sgx_rpm_local_repo || exit 1
 	dnf install -y --nogpgcheck libsgx-launch libsgx-uae-service libsgx-urts libsgx-ae-qve libsgx-dcap-ql libsgx-dcap-ql-devel libsgx-dcap-default-qpl-devel libsgx-dcap-default-qpl || exit 1
 	rm -rf sgx_rpm_local_repo /etc/yum.repos.d/*sgx_rpm_local_repo.repo
-
+elif [ "$OS" == "ubuntu" ]
+then
+# UBUNTU
+       echo 'deb [arch=amd64] https://download.01.org/intel-sgx/sgx_repo/ubuntu/ bionic main' | sudo tee /etc/apt/sources.list.d/intel-sgx.list
+       wget -qO - https://download.01.org/intel-sgx/sgx_repo/ubuntu/intel-sgx-deb.key | sudo apt-key add -
+       apt update
+       apt install -y libsgx-launch libsgx-uae-service libsgx-urts || exit 1
+       apt install -y libsgx-ae-qve libsgx-dcap-ql libsgx-dcap-ql-dev libsgx-dcap-default-qpl-dev libsgx-dcap-default-qpl || exit 1
+fi
 	sed -i "s|PCCS_URL=.*|PCCS_URL=https://$SCS_IP:$SCS_PORT/scs/sgx/certification/v1/|g" /etc/sgx_default_qcnl.conf
 	#Update SCS root CA Certificate in SGX Compute node certificate store in order for  QPL to verify SCS
 	curl -k -H 'Accept:application/x-pem-file' https://$CSP_CMS_IP:$CSP_CMS_PORT/cms/v1/ca-certificates > /etc/pki/ca-trust/source/anchors/skc-lib-cms-ca.cert
